+++ src/common/execution/order-helper.js
@@ -3,12 +3,13 @@ const {
   checkWeb3SecretExists,
   checkRequesterSecretExists,
 } = require('../sms/check');
-const { checkActiveBitInTag } = require('../utils/utils');
+const { checkActiveBitInTag, TAG_MAP } = require('../utils/utils');
 const {
   NULL_ADDRESS,
   NULL_BYTES32,
   IEXEC_REQUEST_PARAMS,
   STORAGE_PROVIDERS,
+  TEE_FRAMEWORKS,
 } = require('../utils/constant');
 const {
   getStorageTokenKeyName,
@@ -18,7 +19,25 @@ const {
   objParamsSchema,
   requestorderSchema,
   throwIfMissing,
+  datasetorderSchema,
+  apporderSchema,
+  tagSchema,
 } = require('../utils/validator');
+const {
+  resolveTeeFrameworkFromApp,
+  showApp,
+} = require('../protocol/registries');
+
+const resolveTeeFrameworkFromTag = async (tag) => {
+  const vTag = await tagSchema().validate(tag);
+  if (checkActiveBitInTag(vTag, TAG_MAP[TEE_FRAMEWORKS.SCONE])) {
+    return TEE_FRAMEWORKS.SCONE;
+  }
+  if (checkActiveBitInTag(vTag, TAG_MAP[TEE_FRAMEWORKS.GRAMINE])) {
+    return TEE_FRAMEWORKS.GRAMINE;
+  }
+  return undefined;
+};
 
 const createObjParams = async ({
   params = {},
@@ -39,7 +58,7 @@ const createObjParams = async ({
   } else {
     inputParams = params;
   }
-  const isTee = checkActiveBitInTag(tag, 1);
+  const isTee = checkActiveBitInTag(tag, TAG_MAP.tee);
   const isCallback = callback !== NULL_ADDRESS;
   const objParams = await objParamsSchema().validate(inputParams, {
     strict: noCast,
@@ -55,22 +74,23 @@ const checkRequestRequirements = async (
   } = throwIfMissing(),
   requestorder = throwIfMissing(),
 ) => {
-  await requestorderSchema().validate(requestorder);
-  const { tag, dataset, callback, params, requester } = requestorder;
+  const vRequestorder = await requestorderSchema().validate(requestorder);
+  const { tag, dataset, callback, params, requester, beneficiary } =
+    vRequestorder;
   const paramsObj = await createObjParams({
     params,
     tag,
     callback,
     noCast: true,
   });
-  const isTee = checkActiveBitInTag(tag, 1);
+  const isTee = checkActiveBitInTag(tag, TAG_MAP.tee);
 
   // check encryption key
   if (paramsObj[IEXEC_REQUEST_PARAMS.IEXEC_RESULT_ENCRYPTION] === true) {
     const isEncryptionKeySet = await checkWeb2SecretExists(
       contracts,
       smsURL,
-      requestorder.beneficiary,
+      beneficiary,
       reservedSecretKeyName.IEXEC_RESULT_ENCRYPTION_PUBLIC_KEY,
     );
     if (!isEncryptionKeySet) {
@@ -135,10 +155,59 @@ const checkRequestRequirements = async (
       ),
     );
   }
-  return true;
+};
+
+const checkDatasetRequirements = async (
+  {
+    contracts = throwIfMissing(),
+    smsURL = throwIfMissing(),
+  } = throwIfMissing(),
+  datasetorder = throwIfMissing(),
+  { tagOverride } = {},
+) => {
+  const vDatasetorder = await datasetorderSchema().validate(datasetorder);
+  const { tag, dataset } = vDatasetorder;
+  const isTee = checkActiveBitInTag(
+    tagOverride ? await tagSchema().validate(tagOverride) : tag,
+    TAG_MAP.tee,
+  );
+  // check tee dataset encryption key
+  if (dataset && dataset !== NULL_ADDRESS && isTee) {
+    const isDatasetSecretSet = await checkWeb3SecretExists(
+      contracts,
+      smsURL,
+      dataset,
+    );
+    if (!isDatasetSecretSet) {
+      throw Error(
+        `Dataset encryption key is not set for dataset ${dataset} in the SMS. Dataset decryption will fail.`,
+      );
+    }
+  }
+};
+
+const checkAppRequirements = async (
+  { contracts = throwIfMissing() } = throwIfMissing(),
+  apporder = throwIfMissing(),
+  { tagOverride } = {},
+) => {
+  const vApporder = await apporderSchema().validate(apporder);
+  const { tag, app } = vApporder;
+  const tagTeeFramework = await resolveTeeFrameworkFromTag(
+    tagOverride ? await tagSchema().validate(tagOverride) : tag,
+  );
+  const appTeeFramework = await showApp(contracts, app).then((res) =>
+    resolveTeeFrameworkFromApp(res.app, { strict: false }),
+  );
+  if (appTeeFramework !== tagTeeFramework) {
+    throw Error('Tag mismatch the TEE framework specified by app');
+  }
 };
 
 module.exports = {
+  resolveTeeFrameworkFromTag,
   createObjParams,
   checkRequestRequirements,
+  checkDatasetRequirements,
+  checkAppRequirements,
 };